Refactor dialog use
